use Coupon API

+++ src/Controller/UserRabattController.php
@@ -89,25 +89,34 @@ class UserRabattController extends AbstractController {
             if (!$jsonAuth->checkJsonCode($data['hash'])) return new Response('Hash Invalid', 403);
             $code = $data["code"];
 
+
             $Rabatt = $this->getDoctrine()->getRepository(UserRabatte::class)->findBy(["Rabatt_Code" => $code]);
             if (count($Rabatt) == 0) return new Response("Coupon Code Not Found. Please try again with another code", 400);
             if (count($Rabatt) >= 2) return new Response("Too many codes found. Please contact the administrator", 400);
             /** @var UserRabatte $Rabatt */
             $Rabatt = $Rabatt[0];
-            if ($Rabatt->getUsed() == true) return new Response("Code has already been used", 400);
-
-            $entityManager->remove($Rabatt);
-            $entityManager->flush();
 
             /** @var Rabatt $Rabatte */
             $Rabatte = $this->getDoctrine()->getRepository(Rabatt::class)->findBy(["id" => $Rabatt->getFKRabattID()]);
+            /** @var Firma $FIRMA */
             $FIRMA = $this->getDoctrine()->getRepository(Firma::class)->findBy(['id' => $Rabatte->getFKFirmaID()])[0];
-            $STATISTIK = new Statistik();
-            $STATISTIK->setDate(new DateTime("0 days ago"));
-            $STATISTIK->setType("gekauft");
-            $STATISTIK->setFKFirmaID($FIRMA->getId());
 
-            return new Response('successful', 200);
+            $RECHTE = $jsonAuth->returnRechteFromHash($data['hash'], $FIRMA->getFirmanname());
+            if ($RECHTE >= 1) {
+                if ($Rabatt->getUsed() == true) return new Response("Code has already been used", 400);
+
+                $entityManager->remove($Rabatt);
+                $entityManager->flush();
+
+                $STATISTIK = new Statistik();
+                $STATISTIK->setDate(new DateTime("0 days ago"));
+                $STATISTIK->setType("gekauft");
+                $STATISTIK->setFKFirmaID($FIRMA->getId());
+                $entityManager->persist($STATISTIK);
+                $entityManager->flush();
+
+                return new Response('successful', 200);
+            }
         } else {
             return new Response("", 404);
         }
